Remove MediaPlayerPage Initialization composition templates and add logic to page templates
