--- templates/Uwp/_composition/CodeBehind/Page.MediaPlayer.AddNavigation_Blank_SplitView/Views/wts.ItemNamePage_postaction.xaml.cs
@@ -1,28 +1,0 @@
-﻿//{[{
-using Windows.UI.Xaml.Navigation;
-//}]}
-
-namespace Param_ItemNamespace.Views
-{
-    public sealed partial class wts.ItemNamePage : Page, INotifyPropertyChanged
-    {
-        public wts.ItemNamePage()
-        {
-        }
-
-        //{[{
-        protected override void OnNavigatedTo(NavigationEventArgs e)
-        {
-            base.OnNavigatedTo(e);
-            mpe.MediaPlayer.PlaybackSession.PlaybackStateChanged += PlaybackSession_PlaybackStateChanged;
-        }
-
-        protected override void OnNavigatedFrom(NavigationEventArgs e)
-        {
-            base.OnNavigatedFrom(e);
-            mpe.MediaPlayer.Pause();
-            mpe.MediaPlayer.PlaybackSession.PlaybackStateChanged -= PlaybackSession_PlaybackStateChanged;
-        }
-        //}]}
-    }
-}
