--- templates/Uwp/_composition/CodeBehind/Page.MediaPlayer.AddNavigation_TabbedPivot/Views/wts.ItemNamePage_searchreplace.xaml.cs
@@ -1,3 +1,0 @@
-﻿public sealed partial class wts.ItemNamePage : Page, INotifyPropertyChanged
-^^^-searchabove-replacebelow-vvv
-public sealed partial class wts.ItemNamePage : Page, INotifyPropertyChanged, IPivotPage
