--- templates/Uwp/_composition/CodeBehind/Page.MediaPlayer.AddNavigation_TabbedPivotVB/Views/wts.ItemNamePage_searchreplace.xaml.vb
@@ -1,3 +1,0 @@
-﻿        Implements INotifyPropertyChanged
-^^^-searchabove-replacebelow-vvv
-        Implements INotifyPropertyChanged, IPivotPage
