--- templates/Uwp/_composition/_shared/Page.MediaPlayer.AddNavigation_Blank_SplitViewVB/.template.config/template.json
@@ -1,28 +1,0 @@
-﻿{
-  "author": "Microsoft Community",
-  "classifications": [
-    "Universal"
-  ],
-  "name": "Page.MediaPlayer.AddNavigation_Blank_SplitView.VB",
-  "tags": {
-    "language": "VisualBasic",
-    "type": "item",
-    "wts.platform": "Uwp",
-    "wts.type": "composition",
-    "wts.version": "1.0.0",
-    "wts.compositionFilter": "$framework == MVVMBasic|MVVMLight & $projectType == Blank|SplitView & identity == wts.Page.MediaPlayer.VB"
-  },
-  "sourceName": "wts.ItemName",
-  "preferNameDirectory": true,
-  "PrimaryOutputs": [],
-  "symbols": {
-    "wts.rootNamespace": {
-      "type": "parameter",
-      "replaces": "Param_RootNamespace"
-    },
-    "wts.itemNamespace": {
-      "type": "parameter",
-      "replaces": "Param_ItemNamespace"
-    }
-  }
-}
