--- templates/Uwp/_composition/_shared/Page.MediaPlayer.AddNavigation_Blank_SplitViewVB/Views/wts.ItemNamePage_postaction.xaml.vb
@@ -1,18 +1,0 @@
-﻿Namespace Views
-    Public NotInheritable Partial Class wts.ItemNamePage
-        Inherits Page
-        '{[{
-
-        Protected Overrides Sub OnNavigatedTo(e As NavigationEventArgs)
-            MyBase.OnNavigatedTo(e)
-            AddHandler mpe.MediaPlayer.PlaybackSession.PlaybackStateChanged, AddressOf PlaybackSession_PlaybackStateChanged
-        End Sub
-
-        Protected Overrides Sub OnNavigatedFrom(e As NavigationEventArgs)
-            MyBase.OnNavigatedFrom(e)
-            mpe.MediaPlayer.Pause()
-            RemoveHandler mpe.MediaPlayer.PlaybackSession.PlaybackStateChanged, AddressOf PlaybackSession_PlaybackStateChanged
-        End Sub
-        '}]}
-    End Class
-End Namespace
