--- templates/Uwp/_composition/_shared/Page.MediaPlayer.AddNavigation_TabbedPivot/Views/wts.ItemNamePage_searchreplace.xaml
@@ -1,3 +1,0 @@
-﻿AutoPlay="True"
-^^^-searchabove-replacebelow-vvv
-AutoPlay="False"
