--- templates/Uwp/_composition/_shared/Page.MediaPlayer.AddNavigation_TabbedPivotVB/Views/wts.ItemNamePage_searchreplace.xaml
@@ -1,3 +1,0 @@
-﻿AutoPlay="True"
-^^^-searchabove-replacebelow-vvv
-AutoPlay="False"
