--- templates/Uwp/_composition/_shared/Page.MediaPlayer.AddNavigation_TabbedPivotVB/Views/wts.ItemNamePage_searchreplace.xaml.vb
@@ -1,4 +1,0 @@
-﻿        Inherits Page
-^^^-searchabove-replacebelow-vvv
-        Inherits Page
-        Implements IPivotPage
